custom-drawers: SerializedPropertyExtension refactoring
