+++ Assets/OneLine/OneLine/Editor/Utils/SerializedPropertyExtension.cs
@@ -63,76 +63,6 @@ namespace OneLine {
             }
         }
 
-        public static T GetCustomAttribute<T>(this SerializedProperty property) where T : Attribute {
-            return GetCustomAttributes<T>(property).FirstOrDefault();
-        }
-
-        public static T[] GetCustomAttributes<T>(this SerializedProperty property) where T : Attribute {
-            return GetCustomAttributes(property, typeof(T))
-                   .Cast<T>()
-                   .ToArray();
-        }
-
-        public static Attribute[] GetCustomAttributes(this SerializedProperty property, Type attributeType = null) {
-            if (attributeType == null) {
-                attributeType = typeof(Attribute);
-            }
-
-
-            var fieldInfo = GetFieldInfo(property);
-            if (fieldInfo == null) {
-                return new Attribute[0];
-            }
-            else {
-                return fieldInfo.GetCustomAttributes(attributeType, false).Cast<Attribute>().ToArray();
-            }
-        }
-
-        public static Type GetRealType(this SerializedProperty property){
-            var fieldInfo = GetFieldInfo(property);
-
-            return fieldInfo == null ? null : fieldInfo.FieldType;
-        }
-
-        public static FieldInfo GetFieldInfo(this SerializedProperty property) {
-            string[] path = property.propertyPath.Split('.');
-
-            bool failed = false;
-            Type type = property.serializedObject.targetObject.GetType();
-            FieldInfo field = null;
-            for (int i = 0; i < path.Length; i++) {
-                field = type.GetField(path[i], BindingFlags.Public 
-                                               | BindingFlags.NonPublic
-                                               | BindingFlags.Instance);
-
-                if (field != null) {
-                    type = field.FieldType;
-                }
-                else {
-                    if (type != typeof(Matrix4x4)){
-                        var message = "[OneLine] Part `{0}` of property path `{1}` doesn't match field definitions of type `{2}`";
-                        Debug.LogWarning(String.Format(message, path[i], property.propertyPath, type.FullName));
-                    }
-
-                    failed = true;
-                    break;
-                }
-
-                int next = i + 1;
-                if (next < path.Length && path[next] == "Array") {
-                    i += 2;
-                    if (type.IsArray) {
-                        type = type.GetElementType();
-                    }
-                    else {
-                        type = type.GetGenericArguments()[0];
-                    }
-                }
-            }
-
-            return failed ? null : field;
-        }
-
         public static bool IsArrayElement(this SerializedProperty property){
             var path = property.propertyPath;
             return path.Substring(path.Length - 1, 1) == "]" ;
